refactor(myweb): move the node/edge renderers into web-graph-renderers
Lift MemberNode and NumberedEdge — with their data shapes (MemberNodeData,
EdgeData), the two interaction contexts WebGraph fills, and the nodeTypes/
edgeTypes maps — out of web-graph.tsx into their own module. The two
renderers mirror each other (each consumes a context the parent provides),
so they read better together. Trims ~150 lines from the graph component;
pure move, no behavior change.
